# Added auto-optimization option for data loaders