--- a/src/super_gradients/__init__.py
+++ b/src/super_gradients/__init__.py
@@ -1,8 +1,8 @@
 from super_gradients.training import ARCHITECTURES, losses, utils, datasets_utils, DataAugmentation, \
-    TestDatasetInterface, SegmentationTestDatasetInterface, DetectionTestDatasetInterface, ClassificationTestDatasetInterface, SgModel
+    TestDatasetInterface, SegmentationTestDatasetInterface, DetectionTestDatasetInterface, \
+    ClassificationTestDatasetInterface, SgModel
 from super_gradients.common import init_trainer, is_distributed
 
 __all__ = ['ARCHITECTURES', 'losses', 'utils', 'datasets_utils', 'DataAugmentation',
            'TestDatasetInterface', 'SgModel', 'SegmentationTestDatasetInterface', 'DetectionTestDatasetInterface',
            'ClassificationTestDatasetInterface', 'init_trainer', 'is_distributed']
-