@avideci 2022-01-19
Did you manage to run it on a machine with running workloads? @shaydeci 
I tested it on CIFAR10 but we need larger datasets as well (ImageNet, COCO...)
We must see if the number will change for a clean environment vs. busy environment with utilization.